+++ crates/cairo-lang-doc/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod db;
+pub mod documentable_item;
 mod markdown;
